requirements.txt
google-adk>=1.8.0
google-genai>=1.9.0
a2a-sdk>=0.3.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
